pyproject.toml
[project]
name = "pyvr"
version = "0.4.1"
description = "GPU-Accelerated Volume Rendering in Python using OpenGL"
authors = [
    {name = "Jixian Li",email = "jixianli@sci.utah.edu"}
]
license = {text = "WTFPL"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
]

[tool.poetry]

[tool.poetry.dependencies]
numpy = ">=2.3"
matplotlib = ">=3.10"
pillow = ">=11.0"
scipy = ">=1.16"
moderngl = ">=5.0"
vtk = "^9.0"

[tool.poetry.group.dev.dependencies]
numpy = ">=2.3"
matplotlib = ">=3.10"
pillow = ">=11.0"
scipy = ">=1.16"
moderngl = ">=5.0"
pytest = ">=7.0"
pytest-cov = ">=7.0"
black = ">=24.9"
isort = ">=6.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
